When a button has no body but does include an icon in its options ({ body: nil, icon: 'rocket' }), it adds an automatic left spacer css class. This was undesirable behavior. There should be no spacing class on an icon when there is no body text to be spaced away from.
This was occurring due to how `capture` collects its content, adding spaces (and fooling the child components into thinking there was a body, and thus to add space).

This is now more explicitly controlled and managed in the NfgUi::Components::Elements::Button. #render now explicitly checks if `body.present?` as well as whether or not an :icon is present for determining what to do with rendering the innards of the component.