Updated environments/test.rb
There seems to be a bug in the Rails 6.0 version that doesn't load
active_support/time in time for configuration files.

See https://github.com/rails/rails/issues/37391

This commit requires 'active_support/time' in the test.rb file in order
to handle the 1.hour method call. We should be able to remove this
require statement in the future.